move ssm validation method to resource
